시퀀스 조회 API 개

+++ api-numbering/src/main/java/com/kko/ohc/numbering/dao/ApiNumberingSequenceDao.java
@@ -2,7 +2,7 @@ package com.kko.ohc.numbering.dao;
 
 public interface ApiNumberingSequenceDao {
 	
-	public int getNext();
+	public double getNext();
 	
 	public int initTable();
